fixing hash thing

+++ Compiler/Parsing/Lexer.cs
@@ -49,7 +49,7 @@ public static class Lexer
     };
     private static readonly char[] _digits = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'];
 
-    public static Token[] LexText(BuildContext ctx, int hash, ReadOnlyMemory<char> text, bool removeUnnecessary = false)
+    public static Token[] LexText(BuildContext ctx, ulong hash, ReadOnlyMemory<char> text, bool removeUnnecessary = false)
     {
         var textSpan = text.Span;
         int index = 0;
